move config structs together
